migrating integration tests from Sashimi to Calamari
